Code cleanup

+++ index.php
@@ -18,7 +18,7 @@ try {
         if ($count > 3) {
             $viewLoaded = $controller->errorCode(400);
         }
-        if (isset($_GET['fetchGrades']) && !isset($_GET['id'])) {
+        if (isset($_GET['fetchGrades']) && !isset($_GET['idSubject'])) {
             if ($_GET['fetchGrades'] == 'true' && $count == 2) {
                 $viewLoaded = $controller->getGradesBySubjects();
             } else if ($_GET['fetchGrades'] == 'false' && $count == 2) {
@@ -26,19 +26,24 @@ try {
             } else {
                 $viewLoaded = $controller->errorCode(400);
             }
-        } else if (isset($_GET['id'])) {
-            if (is_numeric($_GET['id'])) {
+        } else if (isset($_GET['idSubject'])) {
+
+            if (is_numeric($_GET['idSubject'])) {
+                $id = $controller->clear_input($_GET['idSubject']);
                 if (isset($_GET['fetchGrades'])) {
+
                     if ($_GET['fetchGrades'] == 'true') {
-                        $viewLoaded = $controller->getGradesBySubjects($_GET['id']);
+                        $viewLoaded = $controller->getGradesBySubjects($id);
                     } else if ($_GET['fetchGrades'] == 'false') {
-                        $viewLoaded = $controller->getSubject($_GET['id']);
+                        $viewLoaded = $controller->getSubject($id);
                     } else {
                         $viewLoaded = $controller->errorCode(400);
                     }
                 } else if ($count == 2) {
-                    $viewLoaded = $controller->getSubject($_GET['id']);
+                    $viewLoaded = $controller->getSubject($id);
                 }
+            } else {
+                $viewLoaded = $controller->errorCode(400);
             }
         } else if ($count == 1) {
             $viewLoaded = $controller->getSubjects();
@@ -56,53 +61,43 @@ try {
 
         if (isset($_GET['bySubjectId'])) {
             if (is_numeric($_GET['bySubjectId'])) {
-                $viewLoaded = $controller->getGradesFromSubject($_GET['bySubjectId']);
+                $id = $controller->clear_input($_GET['bySubjectId']);
+                $viewLoaded = $controller->getGradesFromSubject($id);
             } else {
                 $viewLoaded = $controller->errorCode(400);
             }
         }
         if (isset($_GET['byGradeId'])) {
             if (is_numeric($_GET['byGradeId'])) {
-                $viewLoaded = $controller->getGrade($_GET['byGradeId']);
+                $id = $controller->clear_input($_GET['byGradeId']);
+                $viewLoaded = $controller->getGrade($id);
             } else {
                 $viewLoaded = $controller->errorCode(400);
             }
         }
-    } else if ($action == 'addGrade' ) {
-        if ($count == 3) {
-            if (isset($_GET['idSubject']) && isset($_GET['value']) && "PUT" == $_SERVER['REQUEST_METHOD']) {
-                $id = $controller->clear_input($_GET['idSubject']);
-                $value = $controller->clear_input($_GET['value']);
-                $viewLoaded = $controller->addGrade($id, $value);
-            } else {
-                $viewLoaded = $controller->errorCode(400);
-            }
+    } else if ($action == 'addGrade') {
+        if (isset($_GET['idSubject']) && isset($_GET['value']) && "PUT" == $_SERVER['REQUEST_METHOD'] && $count == 3) {
+            $id = $controller->clear_input($_GET['idSubject']);
+            $value = $controller->clear_input($_GET['value']);
+            $viewLoaded = $controller->addGrade($id, $value);
         } else {
             $viewLoaded = $controller->errorCode(400);
         }
     } else if ($action == 'deleteGrade') {
-        if ($count == 2) {
+        if (isset($_GET['idGrade']) && "DELETE" == $_SERVER['REQUEST_METHOD'] && $count == 2) {
+
             $id = $controller->clear_input($_GET['idGrade']);
-            if (isset($id)  && "DELETE" == $_SERVER['REQUEST_METHOD']) {
-                $viewLoaded = $controller->deleteGrade($id);
-            } else {
-                $viewLoaded = $controller->errorCode(400);
-            }
+            $viewLoaded = $controller->deleteGrade($id);
         } else {
             $viewLoaded = $controller->errorCode(400);
         }
     } else if ($action == 'subjectAverage') {
-        if ($count == 2) {
-            if (isset($_GET['idSubject']) && is_numeric($_GET['idSubject'])) {
-                $id = $controller->clear_input($_GET['idSubject']);
-                $viewLoaded = $controller->getAverage($id);
-            } else {
-                $viewLoaded = $controller->errorCode(400);
-            }
+        if (isset($_GET['idSubject']) && is_numeric($_GET['idSubject']) && $count == 2) {
+            $id = $controller->clear_input($_GET['idSubject']);
+            $viewLoaded = $controller->getAverage($id);
         } else {
             $viewLoaded = $controller->errorCode(400);
         }
-
     }
     if (!$viewLoaded) {
         $controller->errorCode(404);
